django/Security/requirements.txt
amqp==2.2.2
anyjson==0.3.3
backports-abc==0.5
beautifulsoup4==4.4.1
billiard==3.5.0.3
blessings==1.6
boto==2.38.0
boto3==1.4.4
botocore==1.5.50
bpython==0.16
celery==4.1.0
certifi==2017.1.23
curtsies==0.1.19
Django==1.8.19
django-allauth==0.23.0
django-audiofield==0.8.3
django-celery==3.1.17
django-cors-headers==2.2.0
django-dual-authentication==1.0.0
django-phonenumber-field==0.7.2
django-push-notifications==1.4.1
django-redis-cache==0.13.1
django-redis-sessions==0.5.6
django-rest-auth==0.5.0
django-storages==1.1.8
django-websocket-redis==0.4.7
djangorestframework==3.2.4
docutils==0.13.1
ecdsa==0.13
emoji==0.4.5
future==0.16.0
futures==3.1.1
gevent==1.2.1
greenlet==0.4.11
httpie==0.9.2
jmespath==0.9.2
kombu==4.1.0
lxml==3.4.4
oauthlib==0.7.2
peewee==2.9.1
phonenumbers==7.0.5
Pillow==2.8.2
protobuf==3.2.0
psycopg2==2.7
pycrypto==2.6.1
Pygments==2.0.2
PyJWT==1.1.0
pymediainfo==1.4.0
python-dateutil==2.6.0
python-jose==1.3.2
python-openid==2.2.5
python-social-auth==0.2.7
pytz==2017.2
queries==1.9.1
readline==6.2.4.1
redis==2.10.6
requests==2.6.0
requests-oauthlib==0.4.2
s3transfer==0.1.10
singledispatch==3.4.0.3
six==1.9.0
sleekxmpp==1.3.1
tornado==4.4.2
tornado-redis==2.4.18
uWSGI==2.0.14
vine==1.1.4
wsgiref==0.1.2
